refactor:  create redux logic for cart(create component), add CartEmpty

+++ src/components/CartItem.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import {useDispatch} from "react-redux";
-import {addItem, removeItem} from "../redux/slices/cartSlice";
+import {addItem, minusItem, removeItem} from "../redux/slices/cartSlice";
 
-const CartItem = ({id, type, title, price, count, imageUrl}) => {
+const CartItem = ({id, type, title, price, count, imageUrl, size}) => {
     const dispatch = useDispatch()
 
     const onClickPlus = () => {
@@ -14,8 +14,13 @@ const CartItem = ({id, type, title, price, count, imageUrl}) => {
     }
 
     const onClickMinus = () => {
-        dispatch(removeItem(id))
+        dispatch(minusItem(id))
+    }
 
+    const onClickRemove = () => {
+        if (window.confirm('Ты действительно хочешь удалить товар?')){
+            dispatch(removeItem(id))
+        }
     }
 
     return (
@@ -25,10 +30,10 @@ const CartItem = ({id, type, title, price, count, imageUrl}) => {
             </div>
             <div className="cart__item-info">
                 <h3>{title}</h3>
-                <p>{type}, 26 см.</p>
+                <p>{type}, {size} см.</p>
             </div>
             <div className="cart__item-count">
-                <div className="button button--outline button--circle cart__item-count-minus">
+                <div onClick={onClickMinus} className="button button--outline button--circle cart__item-count-minus">
                     <svg width="10" height="10" viewBox="0 0 10 10" fill="none"
                          xmlns="http://www.w3.org/2000/svg">
                         <path
@@ -41,7 +46,7 @@ const CartItem = ({id, type, title, price, count, imageUrl}) => {
 
                 </div>
                 <b>{count}</b>
-                <div className="button button--outline button--circle cart__item-count-plus">
+                <div onClick={onClickPlus} className="button button--outline button--circle cart__item-count-plus">
                     <svg width="10" height="10" viewBox="0 0 10 10" fill="none"
                          xmlns="http://www.w3.org/2000/svg">
                         <path
@@ -58,7 +63,7 @@ const CartItem = ({id, type, title, price, count, imageUrl}) => {
                 <b>{price * count} ₽</b>
             </div>
             <div className="cart__item-remove">
-                <div className="button button--outline button--circle">
+                <div onClick={onClickRemove} className="button button--outline button--circle">
                     <svg width="10" height="10" viewBox="0 0 10 10" fill="none"
                          xmlns="http://www.w3.org/2000/svg">
                         <path
